+++ ink-wrapper-types/src/lib.rs
@@ -1,98 +1,12 @@
 #[cfg(feature = "aleph_client")]
 mod aleph_client;
+mod calls;
+
 pub mod util;
 
-use std::marker::PhantomData;
-
 use async_trait::async_trait;
+pub use calls::*;
 use ink_primitives::AccountId;
-
-/// Represents a call to a contract constructor.
-#[derive(Debug, Clone)]
-pub struct InstantiateCall<T: Send> {
-    /// The code hash of the contract to instantiate.
-    pub code_hash: [u8; 32],
-    /// The encoded data of the call.
-    pub data: Vec<u8>,
-    /// The salt to use for the contract.
-    pub salt: Vec<u8>,
-    /// A marker for the type of contract to instantiate.
-    _contract: PhantomData<T>,
-}
-
-impl<T: Send> InstantiateCall<T> {
-    /// Create a new instantiate call.
-    pub fn new(code_hash: [u8; 32], data: Vec<u8>) -> Self {
-        Self {
-            code_hash,
-            data,
-            salt: vec![],
-            _contract: Default::default(),
-        }
-    }
-
-    /// Set the salt to use for the instantiation.
-    pub fn with_salt(mut self, salt: Vec<u8>) -> Self {
-        self.salt = salt;
-        self
-    }
-}
-
-/// Represents a mutating contract call to be made.
-#[derive(Debug, Clone)]
-pub struct ExecCall {
-    /// The account id of the contract to call.
-    pub account_id: AccountId,
-    /// The encoded data of the call.
-    pub data: Vec<u8>,
-}
-
-impl ExecCall {
-    /// Create a new exec call.
-    pub fn new(account_id: AccountId, data: Vec<u8>) -> Self {
-        Self { account_id, data }
-    }
-}
-
-/// Represents a read-only contract call to be made.
-#[derive(Debug, Clone)]
-pub struct ReadCall<T: scale::Decode + Send> {
-    /// The account id of the contract to call.
-    pub account_id: AccountId,
-    /// The encoded data of the call.
-    pub data: Vec<u8>,
-    /// A marker for the type to decode the result into.
-    _return_type: PhantomData<T>,
-}
-
-impl<T: scale::Decode + Send> ReadCall<T> {
-    /// Create a new read call.
-    pub fn new(account_id: AccountId, data: Vec<u8>) -> Self {
-        Self {
-            account_id,
-            data,
-            _return_type: Default::default(),
-        }
-    }
-}
-
-/// Represents a call to upload a contract.
-pub struct UploadCall {
-    /// The WASM code to upload.
-    pub wasm: Vec<u8>,
-    /// The expected code hash of the uploaded code.
-    pub expected_code_hash: [u8; 32],
-}
-
-impl UploadCall {
-    /// Create a new upload call.
-    pub fn new(wasm: Vec<u8>, expected_code_hash: [u8; 32]) -> Self {
-        Self {
-            wasm,
-            expected_code_hash,
-        }
-    }
-}
 
 /// A connection with the ability to upload WASM code to the chain.
 #[async_trait]
@@ -110,10 +24,19 @@ pub trait UploadConnection<TxInfo, E>: Sync {
 #[async_trait]
 pub trait SignedConnection<TxInfo, E>: Sync {
     /// Instantiate a contract according to the given `call`.
+    async fn instantiate_tx<T: Send + From<AccountId>>(
+        &self,
+        call: InstantiateCall<T>,
+    ) -> Result<(T, TxInfo), E>;
+
+    /// A convenience method that unpacks the result of `instantiate_tx` if you're not interested in the `TxInfo`.
     async fn instantiate<T: Send + From<AccountId>>(
         &self,
         call: InstantiateCall<T>,
-    ) -> Result<T, E>;
+    ) -> Result<T, E> {
+        let (contract, _) = self.instantiate_tx(call).await?;
+        Ok(contract)
+    }
 
     /// Perform the given mutating call.
     async fn exec(&self, call: ExecCall) -> Result<TxInfo, E>;