+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Ability to send native tokens with calls
+- Ability to get the `tx_info` for a constructor call
+
 ### Fixed
 
 - Invalid version of an indirect dependency causing mangled output when installed without `--locked`